requirements.txt
Pillow==9.5.0
qrcode==6.1
reportlab
